Fixed first & last day reccurance problems
Looked at how this.endDate was actually being used by toString and
stopped updating its date to reflect that (before it was being
interpretted as the date for the end of the first class).

Fixed first day adjustment from start of school to start of class so
that the procedure factored in all the days the class was held on, not
just the first.